Base PlayerController which InputController and AIController inherit from
- also merged CreateBall and GameManager
