+++ Assets/Scripts/GameManager.cs
@@ -4,10 +4,16 @@ using UnityEngine;
 
 public class GameManager : MonoBehaviour
 {
-    public CreateBall createBallScript;
-
     public GameObject player1;
     public GameObject player2;
+
+    public GameObject ballToClone;  // Reference to the object that will be cloned
+    public Vector3 spawnPosition = new Vector3(0, 1.6f, -0.25f);  // Position to spawn the clone
+    public float createBallDelay = 1f;
+
+    private GameObject ball;
+    private bool goCreateBall = false;
+    private float delayTimer = 0;
 
     private int[] score = new int[2];
 
@@ -15,19 +21,51 @@ public class GameManager : MonoBehaviour
     {
         score[0] = 0;
         score[1] = 0;
+
+        createBall();
     }
 
-    // Update is called once per frame
-    void Update()
+    void Update() 
     {
-        
+        if(goCreateBall) 
+        {
+            delayTimer -= Time.deltaTime;
+            if(delayTimer <= 0f) 
+            {
+                // Instantiate (clone) the object at the specified spawn position and default rotation
+                ball = Instantiate(ballToClone, spawnPosition, Quaternion.identity);
+
+                // Check if the cloned object has a Rigidbody, if not add one
+                Rigidbody rb = ball.GetComponent<Rigidbody>();
+                if (rb == null)
+                {
+                    // Add a Rigidbody component if the object doesn't have one
+                    rb = ball.AddComponent<Rigidbody>();
+                }
+
+                goCreateBall = false;
+            }
+        }
     }
 
     public void scorePoint(int player) 
     {
+        destroyBall() ;
         Debug.Log("Player " + (player) + " Scored!");
         ++score[player-1];
         Debug.Log(score[0] + " : " + score[1]);
-        createBallScript.createBall();
+        createBall();
+    }
+
+    private void createBall() 
+    {
+        goCreateBall = true;
+        delayTimer = createBallDelay;
+    }
+
+    private void destroyBall() 
+    {
+        Destroy(ball.gameObject);
+        ball = null;
     }
 }
