+++ Assets/Scripts/InputController.cs
@@ -4,22 +4,30 @@ using UnityEditor.SearchService;
 using UnityEngine;
 using UnityEngine.InputSystem;
 
-public class InputController : MonoBehaviour
+public class InputController : PlayerController
 {
-    public int player = 1;
-    public float speed = 40f;
-    public float rayDistance = 0.5f; // Distance to check for walls
 
     //create private internal references
     private Player1InputActions inputActions1;
     private Player2InputActions inputActions2;
     private InputAction movement;
-    private Vector2 moveInput;     // Stores movement input
-    private Rigidbody rb;
 
-    private void Awake()
+    // Called when movement input is performed
+    void OnMove(InputAction.CallbackContext context)
     {
-        rb = GetComponent<Rigidbody>(); //get rigidbody, responsible for enabling collision with other colliders
+        // Read Vector2 input (WASD or left stick)
+        moveInput = context.ReadValue<Vector2>();
+    }
+
+    // Called when movement input is canceled (released)
+    void OnMoveCanceled(InputAction.CallbackContext context)
+    {
+        // Reset movement input to zero when no input is provided
+        moveInput = Vector2.zero;
+    }
+
+    protected override void DoStart()
+    {
         inputActions1 = new Player1InputActions(); //create new InputActions
         inputActions2 = new Player2InputActions(); //create new InputActions
     }
@@ -43,26 +51,4 @@ public class InputController : MonoBehaviour
         movement.canceled -= OnMoveCanceled;
         movement.Disable();
     }
-
-    // Called when movement input is performed
-    void OnMove(InputAction.CallbackContext context)
-    {
-        // Read Vector2 input (WASD or left stick)
-        moveInput = context.ReadValue<Vector2>();
-    }
-
-    // Called when movement input is canceled (released)
-    void OnMoveCanceled(InputAction.CallbackContext context)
-    {
-        // Reset movement input to zero when no input is provided
-        moveInput = Vector2.zero;
-    }
-
-    private void Update()
-    {
-        Vector3 v3 = new Vector3(0, moveInput.y, 0) * speed * Time.deltaTime; //convert to 3d space
-        Vector3 newPosition = Utils.calculateNewPosition(rb.position, moveInput, speed, rayDistance, GetComponent<Collider>().bounds.extents);
-        rb.MovePosition(newPosition);
-    }
-
 }
