+++ Assets/Scripts/PlayerController.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 07167b02f67d8564481dd90aa460adc3
+guid: 6604856c67fbcf64bbb504b083878d84
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
